Rettelse af Flight Data værdier
Konverteret værdier fra floats (-180000 til 180000) til ints (-180 til
180).
Tilføjet vektor længdeberegning.

+++ AutonomousDrone/src/drone/IDroneControl.java
@@ -34,7 +34,7 @@ public interface IDroneControl {
 
 	boolean isReady();
 
-	float[] getFlightData();
+	int[] getFlightData();
 
 	BufferedImage getbufImg();
 }
